requirements.txt
flask
flask-wtf
email_validator
flask-sqlalchemy
flask-bcrypt
flask-login
flask-mail
itsdangerous==2.0.1
python-dotenv
psycopg2
boto3
psycopg2-binary
stripe
pytz
schedule
boto3